requirements.txt
torch == 1.7.1
pandas == 1.5.3
h5py == 3.8.0
tqdm == 4.64.1
fair-esm == 2.0.0
einops == 0.6.0